requirements.txt
# Core Bot Framework
python-telegram-bot==21.0.1

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# HTTP & WebSockets
aiohttp==3.9.1
websockets==11.0.3
httpx==0.27.0

# Solana & Blockchain
solana==0.34.2
base58==2.1.1

# Data Processing
pandas==2.0.3
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
apscheduler==3.10.4

# Monitoring & Logging
loguru==0.7.2

# PDF Generation (REQUIRED for your reportlab imports)
reportlab==4.0.8

# Additional Required for Your Code
pytz==2023.3
python-dateutil==2.8.2